requirements.txt
boto3>=1.28.0
botocore>=1.31.0
PyYAML>=6.0